refactor(traefik): extract shared helpers and improve readability
- Add  with  and
- Use helpers across parser, validators, and manager
- Minor readability improvements
