+++ src/backends/traefik/templateParser.ts
@@ -122,9 +122,8 @@ export type RenderResult<T> = {
 /**
  * Extract error message from unknown error type.
  */
-export function getErrorMessage(err: unknown): string {
-    return err instanceof Error ? err.message : String(err);
-}
+import { getErrorMessage } from './helpers';
+export { getErrorMessage };
 
 /**
  * Render a template and parse it as YAML.
